pyproject.toml
[build-system]
requires = [ "setuptools>=61.0", "wheel",]
build-backend = "setuptools.build_meta"

[project]
name = "maintsight-pip"
version = "0.5.0"
description = "AI-powered maintenance risk predictor for git repositories with interactive HTML reports and database integration"
readme = "README.md"
requires-python = ">=3.8"
keywords = [ "maintenance", "technical-debt", "risk-prediction", "xgboost", "git", "code-quality", "machine-learning", "repository-health",]
classifiers = [ "Development Status :: 4 - Beta", "Intended Audience :: Developers", "Operating System :: OS Independent", "Programming Language :: Python :: 3", "Programming Language :: Python :: 3.8", "Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Programming Language :: Python :: 3.12", "Topic :: Software Development :: Quality Assurance", "Topic :: Software Development :: Version Control :: Git", "Topic :: Scientific/Engineering :: Artificial Intelligence",]
dependencies = [ "click>=8.0.0", "dill>=0.3.6", "gitpython>=3.1.0", "joblib>=1.3.0", "numpy>=1.21.0", "pandas>=1.3.0", "scikit-learn>=1.1.0", "xgboost>=1.7.0", "rich>=12.0.0", "tqdm>=4.62.0", "typing-extensions>=4.0.0; python_version<'3.10'", "requests>=2.25.0", "python-dotenv>=1.0.0",]
[[project.authors]]
name = "TechDebtGPT Team"
email = "support@techdebtgpt.com"

[[project.maintainers]]
name = "TechDebtGPT Team"
email = "support@techdebtgpt.com"

[project.license]
text = "Apache-2.0"

[project.optional-dependencies]
dev = [ "pytest>=6.0.0", "pytest-cov>=2.12.0", "ruff>=0.1.0", "mypy>=0.950", "pre-commit>=2.15.0",]
api = [ "fastapi>=0.104.0", "uvicorn[standard]>=0.24.0", "pydantic>=2.4.0", "sqlalchemy>=2.0.0", "psycopg2-binary>=2.9.0", "alembic>=1.12.0", "python-dotenv>=1.0.0", "python-multipart>=0.0.6",]

[project.urls]
Homepage = "https://github.com/techdebtgpt/maintsight-pip"
Documentation = "https://github.com/techdebtgpt/maintsight-pip#readme"
Repository = "https://github.com/techdebtgpt/maintsight-pip"
"Bug Tracker" = "https://github.com/techdebtgpt/maintsight-pip/issues"

[project.scripts]
maintsight = "cli:main"
ms = "cli:main"

[tool.setuptools]
packages = [ "services", "models", "utils",]
py-modules = [ "cli", "__init__",]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.setuptools.package-data]
models = [ "*.pkl", "*.json",]
utils = [ "templates/*.html",]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [ "E", "F", "W", "I",]
ignore = [ "E203", "E501", "E402",]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [ "tests",]
